+++ app/Http/Controllers/AccountController.php
@@ -1,7 +1,7 @@
 <?php namespace App\Http\Controllers;
 
 use App\Events\ProfilePictureUpdated;
-use App\User;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Event;
