+++ app/Http/Controllers/Api/BiosController.php
@@ -5,7 +5,7 @@ use Exception;
 use Illuminate\Support\Facades\Auth;
 use App\ApiResources\Bio;
 use App\OAuthGuard\Facades\OAuthGuard;
-use App\User;
+use App\Models\User;
 
 class BiosController extends BaseController
 {
