+++ app/Listeners/UpdateProfilePicture.php
@@ -3,7 +3,7 @@
 namespace App\Listeners;
 
 use App\Events\ProfilePictureUpdated;
-use App\User;
+use App\Models\User;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Support\Facades\File;
