+++ database/seeds/BiosSeeder.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\User;
+use App\Models\User;
 use Illuminate\Database\Seeder;
 
 class BiosSeeder extends Seeder
