+++ database/seeds/TalksSeeder.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\User;
+use App\Models\User;
 use Illuminate\Database\Seeder;
 
 class TalksSeeder extends Seeder
